Move `next_multiple_of()` into `lib.rs`
Move the `next_multiple_of()` filler utility into a more central
location in the crate, now that it's used in multiple places.
