+++ src/lib.rs
@@ -147,6 +147,15 @@ compile_error!(
     "You need to enable at least one of the '2d' or '3d' features for anything to happen."
 );
 
+/// Get the smallest multiple of align greater than or equal to value, where `align` must be a power of two.
+// TODO - filler for usize.next_multiple_of()
+// https://github.com/rust-lang/rust/issues/88581
+pub(crate) fn next_multiple_of(value: usize, align: usize) -> usize {
+    assert!(align & (align - 1) == 0); // power of 2
+    let count = (value + align - 1) / align;
+    count * align
+}
+
 /// Extension trait to write a floating point scalar or vector constant in a
 /// format matching the WGSL grammar.
 ///
@@ -649,6 +658,35 @@ fn gather_removed_effects(
 #[cfg(test)]
 mod tests {
     use super::*;
+
+    const INTS: &[usize] = &[1, 2, 4, 8, 9, 15, 16, 17, 23, 24, 31, 32, 33];
+    const INTS_POW2: &[usize] = &[1, 2, 4, 8, 16, 32, 64, 128, 256, 512, 1024];
+
+    /// Same as `INTS`, rounded up to 16
+    const INTS16: &[usize] = &[16, 16, 16, 16, 16, 16, 16, 32, 32, 32, 32, 32, 48];
+
+    #[test]
+    fn next_multiple() {
+        // align-1 is no-op
+        for &size in INTS {
+            assert_eq!(size, next_multiple_of(size, 1));
+        }
+
+        // zero-sized is always aligned
+        for &align in INTS_POW2 {
+            assert_eq!(0, next_multiple_of(0, align));
+        }
+
+        // size < align : rounds up to align
+        for &size in INTS {
+            assert_eq!(256, next_multiple_of(size, 256));
+        }
+
+        // size > align : actually aligns
+        for (&size, &aligned_size) in INTS.iter().zip(INTS16) {
+            assert_eq!(aligned_size, next_multiple_of(size, 16));
+        }
+    }
 
     #[test]
     fn to_wgsl_f32() {
